Fix: Route generation regression with optionals with default values.
Rails4 and before supported leading segments of urls that were optional
and had default values. This could be used for example for having
locale in the url, and having a global default that was unspecified.
Rails5 with commit 8ca8a2d773b942c4ea broke this behaviour, and commit
0713265fd15f5ce0d6e fixes the case where all optionals have default
values but not the cases when some have it or when some are required.

Fixes #31413